bpsk: move varcode decoding out of demodulator
The varcode routines should not have been inside the demodulator. They
were there as an early proof-of-concept hack.

Return bits from the demodulator when they are discovered, and conduct
the varcode decoding inside the upper `bpsk` layer.

As part of this, a minor change to the PLL nudging was also made -- it
is now nudged in the middle rather than towards the edges, moving it
from being edge-triggered to level-triggered and centering transitions
in the middle of the pll rather than at the ends.

Signed-off-by: Sean Cross <sean@xobs.io>